Removes post-build checks for NSOFF builds.
Removes the attempt to run a post_build script on the platforms
which do not have an NS side being built. This is because the
post_build script is expecting artifacts which are only built in
the NS case, which has been causing failures on the CI when it could
not find those artifacts.

Signed-off-by: Matthew Dalzell <matthew.dalzell@arm.com>
Change-Id: Ia8cabb9fdfab4dc329bf834d403b29a56d182ff6