Eclair-builds: Modify git commands to checkout index and clean
Use git checkout-index to checkout the index into the working
tree (forced, for all files, quiet). Expand on using git clean
to remove all files in the working tree which are not in the
index (forced, for all files and directories, quiet). This
should fixed the recent issue with eclair-daily builds due to
the new RPI2350 patch on psa-arch-tests fetch_repo().

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I85c325fcaa1363d76c7fccf8a197efa58e1e9dab