Checkpatch: Remove the option to update checkpatch from network
checkpatch.pl and related files are now hosted directly in the
CI repo, so remove the step from the Groovy script to require
the update of the files from network (those are now handled
offline when required). Also remove the capability itself from
the run-checkpatch.sh script, which is already deprecated.

Signed-off-by: Antonio de Angelis <antonio.deangelis@arm.com>
Change-Id: I9f7db56ffc0ff590517bc9e559b00593a49f428a