RSE: Add ROM SRAM copy build config
Add a new build config which builds the RSE using the MinSizeRel
CMAKE_BUILD_TYPE and specifies that the RSE should copy the ROM code to
SRAM and execute it there.

Note that we can only use the MinSizeRel build type when using this
config flag, as this requires the BL1_1 code to be smaller than the SRAM
size.

Change-Id: Ie5c7e4a4997767dcdf9ef4e192954bed8f4fc37f
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>