RSE: Add symmetric provisioning build config
Change the existing RSE_PROVISIONING_ASYMMETRIC build config to instead
set the flag to ON. This reflects a change in TF-M which sets it
to OFF by default. We can also remove the build exclusion for the
BL1_1 tests as part of this change as it is no longer required. Also update RSE related builds to use GCC 13.2.

Change-Id: I83e5efd8587a002efb31df58627fa8f464b6b6d0
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>