RSE: Enable CONFIG_TFM_WARNINGS_ARE_ERRORS for base debug build
This allows us to pick-up if any new warnings have been added as part of
a patch. Note that it is only done for the base debug build as that is
currently the only build which is without warnings.

This patch also fixes an issue with the comment of the build name when
the tests are enabled in PCI mode.

Change-Id: Ic7e8a2efe1a8779de27831a71b92b98f14fde55b
Signed-off-by: Jackson Cooper-Driver <jackson.cooper-driver@arm.com>